M-A3 P0.3: split the registry + pure evaluator into Parsek.dll
Move InvariantRegistry into Parsek.dll (Source/Parsek/Analyzer/, namespace
Parsek.Analyzer) and extract a new pure InvariantEvaluator alongside it
(Evaluate(model), Evaluate(model,rules), DiscoverSubjectSchemaGeneration) - the
exact code path the in-game H5 category will reuse. Neither reads a file.

The offline orchestrator (load save dir -> evaluate -> write reports) stays in
Parsek.Tests as a new Analyzer.cs, since it depends on the Tests-only
SaveDirectoryLoader + ReportWriter. It calls InvariantEvaluator.Evaluate.

Analyzer.Evaluate is kept here as a transitional forwarder so this commit lands
the split without touching the ~10 call sites; P0.4 renames them to
InvariantEvaluator.Evaluate / OfflineAnalyzer.Run and drops the shim.
